requirements.txt
streamlit
pycryptodome
python-docx
